requirements/train_pipeline.txt
# Core ML
numpy
pandas
matplotlib
tqdm
pillow
scikit-learn
opencv-python>=4.8
# Tracking
mlflow>=2.12.0
regex>=2024.4.16
# DB/IO helpers
psycopg2-binary
boto3
tensorboard>=2.16
# Optional but handy
pyyaml
seaborn